requirements.txt
Django==1.11.5
django-cors-headers==2.1.0
django-filter==1.0.4
django-sslserver==0.20
djangorestframework==3.6.4
numpy==1.13.3
PyBrain3==3.0.4
pytz==2017.3
scikit-learn==0.19.1
scipy==1.0.0
